move redirector out of controller, add specs
